change indices type int to size_t

+++ filters/include/pcl/filters/uniform_sampling.h
@@ -111,7 +111,7 @@ namespace pcl
       struct Leaf
       {
         Leaf () : idx (-1) { }
-        int idx;
+        size_t idx;                   // index can be larger than int size
       };
 
       /** \brief The 3D grid leaves. */
